+++ src/Services/SpamSubmissionsService.php
@@ -11,6 +11,7 @@
 
 namespace Solspace\Freeform\Services;
 
+use craft\db\Query;
 use craft\records\Element;
 use Solspace\Freeform\Elements\SpamSubmission;
 use Solspace\Freeform\Elements\Submission;
@@ -20,10 +21,14 @@ use Solspace\Freeform\Library\Exceptions\FreeformException;
 
 class SpamSubmissionsService extends SubmissionsService implements SpamSubmissionHandlerInterface
 {
+    const MIN_PURGE_AGE   = 3;
+    const PURGE_CACHE_KEY = 'freeform_purge_spam_cache_key';
+
     /**
      * @inheritdoc
      */
-    public function getSubmissionById($id) {
+    public function getSubmissionById($id)
+    {
         return SpamSubmission::find()->id($id)->one();
     }
 
@@ -42,15 +47,15 @@ class SpamSubmissionsService extends SubmissionsService implements SpamSubmissio
 
         //HACK: this is dirty, but I wasn't able to find better way to
         //      quickly convert SpamSubmission to Submission
-        $element = Element::findOne($submission->id);
+        $element       = Element::findOne($submission->id);
         $element->type = Submission::class;
         $element->save(false);
 
         $layout            = $submission->getForm()->getLayout();
         $integrationsQueue = Freeform::getInstance()->integrationsQueue;
         $tasks             = $integrationsQueue->getTasksBySubmissionId($submission->id);
-        $fields            = array();
-        foreach($tasks as $task) {
+        $fields            = [];
+        foreach ($tasks as $task) {
             $fields[] = $layout->getFieldByHash($task->fieldHash);
         }
 
@@ -64,14 +69,48 @@ class SpamSubmissionsService extends SubmissionsService implements SpamSubmissio
     /**
      * Processes spam submission so it could be processed normally in case of whitelisting
      *
-     * @param Submission $submission
+     * @param Submission      $submission
      * @param AbstractField[] $mailingListOptedInFields
      */
-    public function postProcessSubmission(Submission $submission, array $mailingListOptedInFields) {
-         if (!$submission instanceof SpamSubmission || !$submission->id) {
+    public function postProcessSubmission(Submission $submission, array $mailingListOptedInFields)
+    {
+        if (!$submission instanceof SpamSubmission || !$submission->id) {
             throw new FreeformException('Invalid $submission, can process only stored SpamSubmission instances.');
         }
 
         Freeform::getInstance()->integrationsQueue->enqueueIntegrations($submission, $mailingListOptedInFields);
+    }
+
+    /**
+     * Removes all old spam submissions according to the spam age set in settings
+     */
+    public function purgeSubmissions()
+    {
+        $hasBeenPurgedRecently = \Craft::$app->cache->get(static::PURGE_CACHE_KEY);
+        if ($hasBeenPurgedRecently) {
+            return;
+        }
+
+        $age = Freeform::getInstance()->settings->getPurgableSpamAgeInDays();
+        if (\is_int($age) && $age >= static::MIN_PURGE_AGE) {
+            $date = new \DateTime("-$age days");
+
+            $ids = (new Query())
+                ->select(['id'])
+                ->from(Submission::TABLE)
+                ->where(['<', 'dateCreated', $date->format('Y-m-d H:i:s')])
+                ->andWhere(['isSpam' => true])
+                ->column();
+
+            \Craft::$app->db
+                ->createCommand()
+                ->delete(
+                    Element::tableName(),
+                    ['id' => $ids]
+                )
+                ->execute();
+
+            \Craft::$app->cache->set(static::PURGE_CACHE_KEY, true, static::PURGE_CACHE_TTL);
+        }
     }
 }